requirements/common.txt
Django==1.10.0
djangorestframework==3.4.1
python-gnupg==0.3.8
requests==2.11.0
django-ckeditor==5.1.1
TwitterAPI==2.4.2
django-extra-views==0.8.0
Pillow==3.4.2
